refactor(gateway): finalize kernel composition
